services/ingestion/requirements.txt
psycopg2-binary==2.9.9
httpx==0.27.0
feedparser==6.0.11
beautifulsoup4==4.12.3
trafilatura==1.8.1
apscheduler==3.10.4
pydantic==2.6.4
pydantic-settings==2.2.1
python-dateutil==2.9.0
certifi==2024.2.2
